add support for player build assemblies

+++ Core/Pipelines/Jobs/StageAssemblies.cs
@@ -24,6 +24,7 @@ namespace ThunderKit.Core.Pipelines.Jobs
         void CopyReferences(AssemblyDefinitionAsset[] assemblyDefs, string ouptputPath)
         {
             var scriptAssemblies = Path.Combine("Library", "ScriptAssemblies");
+            var playerScriptAssemblies = Path.Combine("Library", "PlayerScriptAssemblies");
 
             foreach (var definition in assemblyDefs)
             {
@@ -33,25 +34,27 @@ namespace ThunderKit.Core.Pipelines.Jobs
                 Directory.CreateDirectory(Path.GetDirectoryName(fileOutputBase));
 
                 if (File.Exists($"{fileOutputBase}.dll")) File.Delete($"{fileOutputBase}.dll");
-                File.Copy(Path.Combine(scriptAssemblies, $"{assemblyDef.name}.dll"), $"{fileOutputBase}.dll");
+                var playerBuildExists = File.Exists(Path.Combine(playerScriptAssemblies, $"{assemblyDef.name}.dll"));
+                if (playerBuildExists)
+                    File.Copy(Path.Combine(playerScriptAssemblies, $"{assemblyDef.name}.dll"), $"{fileOutputBase}.dll");
+                else
+                    File.Copy(Path.Combine(scriptAssemblies, $"{assemblyDef.name}.dll"), $"{fileOutputBase}.dll");
 
                 if (stageDebugDatabases)
                 {
-                    string monodatabase = $"{Path.Combine(scriptAssemblies, fileName)}.dll.mdb";
-                    if (File.Exists(monodatabase))
-                    {
-                        if (File.Exists($"{fileOutputBase}.dll.mdb")) File.Delete($"{fileOutputBase}.dll.mdb");
+                    if (File.Exists($"{fileOutputBase}.dll.mdb")) File.Delete($"{fileOutputBase}.dll.mdb");
+                    if (File.Exists($"{fileOutputBase}.pdb")) File.Delete($"{fileOutputBase}.pdb");
 
-                        File.Copy(monodatabase, $"{fileOutputBase}.dll.mdb");
-                    }
+                    string pdbPath = Path.Combine(scriptAssemblies, $"{fileName}.pdb");
+                    string playerPdbPath = Path.Combine(playerScriptAssemblies, $"{fileName}.pdb");
+                    string mdbPath = $"{Path.Combine(scriptAssemblies, fileName)}.dll.mdb";
+                    string playerMdbPath = $"{Path.Combine(playerScriptAssemblies, fileName)}.dll.mdb";
 
-                    string programdatabase = Path.Combine(scriptAssemblies, $"{fileName}.pdb");
-                    if (File.Exists(programdatabase))
-                    {
-                        if (File.Exists($"{fileOutputBase}.pdb")) File.Delete($"{fileOutputBase}.pdb");
+                    if (File.Exists(playerMdbPath)) File.Copy($"{Path.Combine(playerScriptAssemblies, fileName)}.dll.mdb", $"{fileOutputBase}.dll.mdb");
+                    else if (File.Exists(mdbPath)) File.Copy(mdbPath, $"{fileOutputBase}.dll.mdb");
 
-                        File.Copy(programdatabase, $"{fileOutputBase}.pdb");
-                    }
+                    if (File.Exists(playerPdbPath)) File.Copy($"{Path.Combine(playerScriptAssemblies, fileName)}.dll.pdb", $"{fileOutputBase}.dll.pdb");
+                    else if (File.Exists(pdbPath)) File.Copy(pdbPath, $"{fileOutputBase}.dll.pdb");
                 }
             }
         }
